+++ Dropdown/DropdownList.js
@@ -7,8 +7,7 @@ import IdProvider from '@enact/ui/internal/IdProvider';
 import ri from '@enact/ui/resolution';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
-import {Component} from 'react';
-import ReactDOM from 'react-dom';
+import {Component, createRef} from 'react';
 
 import $L from '../internal/$L';
 import Icon from '../Icon';
@@ -207,6 +206,7 @@ const DropdownListSpotlightDecorator = hoc((config, Wrapped) => {
 		constructor (props) {
 			super(props);
 
+			this.clientSiblingRef = createRef(null);
 			this.state = {
 				prevChildren: props.children,
 				prevFocused: null,
@@ -220,13 +220,6 @@ const DropdownListSpotlightDecorator = hoc((config, Wrapped) => {
 			if (this.props.handleSpotlightPause) {
 				this.props.handleSpotlightPause(false);
 			}
-
-			// eslint-disable-next-line react/no-find-dom-node
-			this.node = ReactDOM.findDOMNode(this);
-			Spotlight.set(this.node.dataset.spotlightId, {
-				defaultElement: '[data-selected="true"]',
-				enterTo: 'default-element'
-			});
 		}
 
 		componentDidUpdate () {
@@ -296,8 +289,9 @@ const DropdownListSpotlightDecorator = hoc((config, Wrapped) => {
 
 		handleFocus = (ev) => {
 			const current = ev.target;
+			const dropdownListNode = this.clientSiblingRef?.current?.previousElementSibling;
 			if (this.state.ready === ReadyState.DONE && !Spotlight.getPointerMode() &&
-				current.dataset['index'] != null && this.node.contains(current)
+				current.dataset['index'] != null && dropdownListNode.contains(current)
 			) {
 				const focusedIndex = Number(current.dataset['index']);
 				const lastFocusedKey = getKey({children: this.props.children, selected: focusedIndex});
@@ -314,7 +308,10 @@ const DropdownListSpotlightDecorator = hoc((config, Wrapped) => {
 			delete props.handleSpotlightPause;
 
 			return (
-				<Wrapped {...props} onFocus={this.handleFocus} scrollTo={this.setScrollTo} />
+				<>
+					<Wrapped {...props} onFocus={this.handleFocus} scrollTo={this.setScrollTo} />
+					<div style={{display: 'none'}} ref={this.clientSiblingRef} />
+				</>
 			);
 		}
 	};